% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{course_copy}
\alias{course_copy}
\title{course_copy}
\usage{
course_copy(course_id, source_course_id)

course_copy(course_id, source_course_id)
}
\arguments{
\item{course_id}{The course id}

\item{source_course_id}{The source course id}
}
\value{
a content migration.

a content migration.
}
\description{
This function creates a standard course copy from one course to another.

This function creates a standard course copy from one course to another.
}
