% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{create_assignment}
\alias{create_assignment}
\title{Create assignment}
\usage{
create_assignment(
  course_id,
  assignment_name,
  position = NULL,
  submission_type = "none",
  notify_of_update = F,
  points_possible = 10,
  grading_type = "points",
  published = F
)
}
\arguments{
\item{course_id}{canvas id of the course (integer)}

\item{assignment_name}{name of the assignment (string)}

\item{position}{position of the assignment in the list of assignments. Default to 0 -> end of the list (integer)}

\item{submission_type}{online_quiz/none/on_paper/discussion_topic/external_tool (string)}

\item{notify_of_update}{notifies student of changes if TRUE (boolean)}

\item{points_possible}{max points possible, default is 10 (integer)}

\item{grading_type}{pass_fail/percent/letter_grade/gpa_scale/points/not_graded (string)}

\item{published}{is the assignment published, default is False (boolean)}
}
\value{
server response. Either 201 status code if everything went correctly or a specific http status warning.
}
\description{
Create a new assignment in course
}
