% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{create_course}
\alias{create_course}
\title{Create course}
\usage{
create_course(
  account_id,
  course_name,
  course_code,
  restrict_enrollments_to_course_dates = TRUE,
  term_id
)
}
\arguments{
\item{account_id}{the id of the account in which you want to create the course}

\item{course_name}{the name of the new (string)}

\item{course_code}{the course code/reference code for the course (string)}

\item{restrict_enrollments_to_course_dates}{set to true to restrict user enrollments to the start and end dates of the course}

\item{term_id}{the ID of the term to create the course in (integer)}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
Create a new course
}
