% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{create_course_group_category}
\alias{create_course_group_category}
\title{create_course_group_category}
\usage{
create_course_group_category(course_id, groupcat_name,
  self_signup = NULL, auto_leader = NULL, group_limit = NULL,
  create_group_count = NULL)
}
\arguments{
\item{create_group_count}{}
}
\description{
create_course_group_category
}
