% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{create_course_group_in_category}
\alias{create_course_group_in_category}
\title{create_course_group_in_category}
\usage{
create_course_group_in_category(groupcat_id, name, description = NULL,
  is_public = NULL, join_level = NULL, storage_quota_mb = NULL,
  sis_group_id = NULL)
}
\arguments{
\item{groupcat_id}{the canvas id of the group category (integer)}

\item{name}{the name of the group(string/character)}

\item{description}{a description of the group (string/character)}

\item{is_public}{make the group public (boolean)}

\item{storage_quota_mb}{storage quota of the group page in megabytes (integer)}

\item{sis_group_id}{the sis id of the group (integer)}
}
\description{
Creates a group in a predefined Canvas group category
}
