% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{create_group_membership}
\alias{create_group_membership}
\title{create_group_membership}
\usage{
create_group_membership(group_id, user_id)
}
\arguments{
\item{user_id}{}
}
\description{
create_group_membership
}
