% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{create_section_in_course}
\alias{create_section_in_course}
\title{Create a section in a course}
\usage{
create_section_in_course(
  course_id,
  name,
  sis_section_id = NULL,
  integration_id = NULL,
  start_at = NULL,
  end_at = NULL,
  restrict_enrollments_to_section_dates = NULL,
  enable_sis_reactivation = NULL
)

create_section_in_course(
  course_id,
  name,
  sis_section_id = NULL,
  integration_id = NULL,
  start_at = NULL,
  end_at = NULL,
  restrict_enrollments_to_section_dates = NULL,
  enable_sis_reactivation = NULL
)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{name}{the name of the section (string/character)}

\item{sis_section_id}{the sis id of the section (integer)}

\item{integration_id}{}

\item{start_at}{start date of the section}

\item{end_at}{end date of the section}

\item{restrict_enrollments_to_section_dates}{restrict enrollments to a specific date}

\item{enable_sis_reactivation}{}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.

server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
This function creates a section in a Canvas course

This function creates a section in a Canvas course
}
\examples{

create_section_in_course(course_id = 6824, name = "Werkgroep_1")

create_section_in_course(course_id = 6824, name = "Werkgroep_1")
}
