% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{crosslist_section_to_course}
\alias{crosslist_section_to_course}
\title{Crosslist section}
\usage{
crosslist_section_to_course(section_id, new_course_id)
}
\arguments{
\item{section_id}{the canvas id of the section (integer)}

\item{new_course_id}{the canvas id of the new course (integer)}
}
\description{
This functions crosslist a section to a new course
}
