% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{delete_course_module}
\alias{delete_course_module}
\title{delete_course_module
Deletes a specified module in a course.}
\usage{
delete_course_module(course_id, module_id)
}
\arguments{
\item{module_id}{}
}
\description{
delete_course_module
Deletes a specified module in a course.
}
