% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{delete_course_page}
\alias{delete_course_page}
\title{delete_course_page
Deletes a specific page in a course. Does not delete pages pushed by blue_print courses}
\usage{
delete_course_page(course_id, html_url)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{html_url}{the html_url of the page. Can be found in the get_course_pages return. (character)}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
delete_course_page
Deletes a specific page in a course. Does not delete pages pushed by blue_print courses
}
