% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{delete_course_section}
\alias{delete_course_section}
\title{Delete course section}
\usage{
delete_course_section(section_id)

delete_course_section(section_id)
}
\arguments{
\item{section_id}{the canvas id of the section (integer)}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.

server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
Deletes a specified section in a course.
 
PLEASE NOTE: To delete a course section, the section needs to be empty, This function does not empty the section. Use delete_sections for this task.

Deletes a specified section in a course.
 
PLEASE NOTE: To delete a course section, the section needs to be empty, This function does not empty the section. Use delete_sections for this task.
}
\examples{

delete_course_section(12345)

delete_course_section(12345)
}
