% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{delete_section}
\alias{delete_section}
\title{delete_course_section
Deletes a specified module in a course.}
\usage{
delete_section(section_id)
}
\arguments{
\item{section_id}{the canvas id of the section (integer)}
}
\description{
delete_course_section
Deletes a specified module in a course.
}
