% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{delete_section_enrollments}
\alias{delete_section_enrollments}
\title{delete_section_enrollments}
\usage{
delete_section_enrollments(course_id, enrollment_id, task)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{task}{action to take on the enrollment}

\item{section_id}{the canvas id of the section (integer)}
}
\description{
Deletes a specified section_enrollment from a section in a course. This does not mean a user will be 
unenrolled from the course, only from the section. You are able to either conclude, delete, inactivate or
deactivate the enrollment.
}
