% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_sections.R
\name{delete_sections}
\alias{delete_sections}
\title{Delete sections from a course}
\usage{
delete_sections(course_id, section_ids = "all")
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{section_ids}{all by default or specify by entering canvas ids of the sections}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
This function allows to completely delete user created sections from a course. SIS created section will remain intact.
The function first empties a section by deleting all section enrollments and then removing the section. 
Please note that this will not delete a user from a course.
The function allows to either delete all sections from a course by default, or a specified id's
}
\examples{

delete_sections(6348)
}
