% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_grades.R
\name{download_grades}
\alias{download_grades}
\title{Download Grades}
\usage{
download_grades(course_id, csv_export = F)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{csv_export}{indicate if you want a csv export to be saved (boolean)}
}
\value{
a dataframe and/or csv export of the gradebook
}
\description{
This functions downloads grades on assignments from the gradebook of a specific course
}
\examples{

download_grades(6348, csv_export = F)

}
