% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_quiz_results.R
\name{download_quiz_results}
\alias{download_quiz_results}
\title{Download Grades}
\usage{
download_quiz_results(course_id, quiz_id, csv_export = F)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{quiz_id}{the canvas id of the quiz (integer)}

\item{csv_export}{indicate if you want a csv export to be saved (boolean)}
}
\value{
a dataframe and/or csv export of the gradebook
}
\description{
This functions downloads results from a specific quiz and specific course
}
\examples{

download_grades(6348, 1234, csv_export = F)

}
