% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{enroll_user_in_course}
\alias{enroll_user_in_course}
\title{Enroll a User in a Course}
\usage{
enroll_user_in_course(
  course_id,
  user_id,
  enrollment_type = NULL,
  enrollment_state = NULL,
  course_section_id = NULL,
  interaction_limit = FALSE
)

enroll_user_in_course(
  course_id,
  user_id,
  enrollment_type = NULL,
  enrollment_state = NULL,
  course_section_id = NULL,
  interaction_limit = FALSE
)
}
\arguments{
\item{course_id}{the course id (integer)}

\item{user_id}{the canvas id of the user (string). Alternatively, the sis_id can
be used by setting "sis_login_id:" in front of the sis_id.}

\item{enrollment_type}{the type of the enrollment (string). Allowed values are "StudentEnrolment",
"TeacherEnrollment", "TaEnrollment", ObserverEnrollment" and "DesignerEnrollment".}

\item{enrollment_state}{the state of the enrollment (string). Allowed values are "active", "invited"
and "inactive". Default is invited. N.B. if this parameter is incorrectly specified,
the default is used.}

\item{course_section_id}{the id of the section for the enrollment. Can be located in the URL
when visiting the section page in the browser.}

\item{interaction_limit}{limit interaction with users to section only.}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.

server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
This function is used to enroll a user in a specific course.
Enrollment-settings can be altered, such as silent or invited enrollment and
teacher or student enrollment.

This function is used to enroll a user in a specific course.
Enrollment-settings can be altered, such as silent or invited enrollment and
teacher or student enrollment.
}
\examples{

enroll_user_in_course(course_id = 1867, user_id = "sis_login_id:pbosman", enrollment_type = "TeacherEnrollment", enrollment_state = "inactive")


enroll_user_in_course(course_id = 1867, user_id = "sis_login_id:pbosman", enrollment_type = "TeacherEnrollment", enrollment_state = "inactive")

}
