% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_assignmentgroups_in_course}
\alias{get_assignmentgroups_in_course}
\title{Get assignment groups in course}
\usage{
get_assignmentgroups_in_course(course_id)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}
}
\value{
A dataframe containing the assignment group information from the specified course
}
\description{
Get all assignment groups in Course
}
