% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{get_assignments_in_course}
\alias{get_assignments_in_course}
\title{Get assignments in course}
\usage{
get_assignments_in_course(course_id)

get_assignments_in_course(course_id)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}
}
\value{
A dataframe containing assignment information of all assignments from the specified course

A dataframe containing assignment information of all assignments from the specified course
}
\description{
Get all assignments information in Course

Get all assignments information in Course
}
\examples{

get_assignments_in_course(6824)

get_assignments_in_course(6824)
}
