% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint_functions.R
\name{get_blueprint_info}
\alias{get_blueprint_info}
\title{Get Blueprint Information}
\usage{
get_blueprint_info(course_id, template_id = "default")
}
\arguments{
\item{course_id}{the canvas id of the course_id (integer)}

\item{template_id}{Using 'default' as the template_id should suffice for the current implmentation (as there should be only one template per course). However, using specific template ids may become necessary in the future}
}
\value{
a dataframe containing information about the Blueprint
}
\description{
Get information about the blueprint page
}
