% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_course_info}
\alias{get_course_info}
\title{Get course information}
\usage{
get_course_info(course_id, include = NULL)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}
}
\value{
A dataframe containing information from the specified course
}
\description{
Get course information
}
\examples{

get_course_info(6824)
}
