% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_course_modules}
\alias{get_course_modules}
\title{get_course_modules
Returns a data frame with all modules of a selected course.}
\usage{
get_course_modules(course_id)
}
\arguments{
\item{course_id}{}
}
\description{
get_course_modules
Returns a data frame with all modules of a selected course.
}
