% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_course_pages}
\alias{get_course_pages}
\title{get_course_pages
Returns a data frame with all the pages in a course.}
\usage{
get_course_pages(course_id)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}
}
\value{
A dataframe containing course page information.
}
\description{
get_course_pages
Returns a data frame with all the pages in a course.
}
\examples{

get_course_pages(course_id = 6824)
}
