% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{get_course_rubric}
\alias{get_course_rubric}
\title{Get Course Rubric}
\usage{
get_course_rubric(course_id, rubric_id)

get_course_rubric(course_id, rubric_id)
}
\arguments{
\item{course_id}{the canvas id of the course_id (integer)}

\item{rubric_id}{the canvas id of the assignment id (integer)}
}
\value{
dataframe with course info

dataframe with course info
}
\description{
Get information about a rubric

Get information about a rubric
}
