% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{get_course_sections}
\alias{get_course_sections}
\title{get_course_sections
Get sections information in course}
\usage{
get_course_sections(course_id)

get_course_sections(course_id)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}
}
\value{
A dataframe containing information about section in the specified course.

A dataframe containing information about section in the specified course.
}
\description{
get_course_sections
Get sections information in course

get_course_sections
Get sections information in course
}
\examples{

get_course_sections(course_id = 6824)

get_course_sections(course_id = 6824)
}
