% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_course_tabs}
\alias{get_course_tabs}
\title{Get course tabs}
\usage{
get_course_tabs(course_id)
}
\arguments{
\item{course_id}{the canvas id of the course_id (integer)}
}
\value{
a paginated list
}
\description{
Get tabs from course
}
