% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_courses_in_account}
\alias{get_courses_in_account}
\title{Get all Courses in an Account}
\usage{
get_courses_in_account(account_id, include = NULL)
}
\arguments{
\item{account_id}{the account id (integer)}

\item{include}{"syllabus_body", "term", "course_progress", "storage_quota_used_mb", "total_students", "teachers", "account_name" (character)}
}
\value{
a data frame of active courses in account
}
\description{
This functions gets all the courses in a specific account. It includes information such
as course id, sis id and the id of the term. You can get extra information like syllabus_body, term, course_progress, storage_quota_used_mb, total_students, teachers, account_name
via the include parameter.
}
\examples{
get_courses_in_account(123)

}
