% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_enrollment_terms}
\alias{get_enrollment_terms}
\title{Get enrollment terms}
\usage{
get_enrollment_terms(account_id = 1)
}
\arguments{
\item{account_id}{the canvas id of the course_id (integer)}
}
\value{
dataframe with enrollment term info
}
\description{
List all enrollment terms
}
