% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_groups_in_course}
\alias{get_groups_in_course}
\title{Get groups in course}
\usage{
get_groups_in_course(course_id)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}
}
\description{
Get groups in course
}
