% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_pages_in_course}
\alias{get_pages_in_course}
\title{Get pages in course}
\usage{
get_pages_in_course(course_id)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}
}
\description{
Get all page information in course
}
\examples{

get_pages_in_course(6824)
}
