% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_quizzes_in_course}
\alias{get_quizzes_in_course}
\title{Get quizzes in course}
\usage{
get_quizzes_in_course(course_id)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}
}
\value{
A dataframe containing quiz information of all quizzes from the specified course
}
\description{
Get all quizzes information in Course
}
\examples{

get_assignments_in_course(6824)
}
