% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_section_enrollments}
\alias{get_section_enrollments}
\title{get_section_enrollments}
\usage{
get_section_enrollments(section_id)
}
\arguments{
\item{section_id}{}
}
\description{
get_section_enrollments
}
