% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_users_in_account}
\alias{get_users_in_account}
\title{Get users in account}
\usage{
get_users_in_account(account_id, enrollment_type = NULL)
}
\arguments{
\item{account_id}{the canvas id of the account (integer)}

\item{enrollment_type}{the canvas enrollment type. Use either "student", "teacher", "ta", "observer" or "designer" (character)}
}
\description{
Get users in account
}
