% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_users_in_course}
\alias{get_users_in_course}
\title{Get users in course}
\usage{
get_users_in_course(course_id, enrollment_type = NULL,
  enrollment_state = NULL)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{enrollment_type}{the canvas enrollment type to return. Use either "StudentEnrollment", "TeacherEnrollment", "TaEnrollment", "ObserverEnrollment" or "DesignerEnrollment". Default returns all roles. (character)}

\item{enrollment_state}{include users with a specific state. Use either "active", "invited", "deleted" or "completed". Default is "active"  and "invited" (character)}
}
\description{
Get users in course
}
