% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{grade_assignment}
\alias{grade_assignment}
\title{Grade Assignment}
\usage{
grade_assignment(course_id, assignment_id, user_id, grade = NULL, excuse = F)
}
\arguments{
\item{course_id}{the canvas id of the course_id (integer)}

\item{assignment_id}{the canvas id of the assignment id (integer)}

\item{user_id}{the canvas id of the user (integer)}

\item{grade}{the grade example "complete" (string)}

\item{excuse}{excuse}
}
\value{
a server status report
}
\description{
Grade Assignment
}
