% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{list_course_files}
\alias{list_course_files}
\title{List course files}
\usage{
list_course_files(course_id, include = NULL)
}
\arguments{
\item{course_id}{The canvas id of the course (integer)}

\item{include}{Include user information to the data. (character)}
}
\value{
A dataframe with information about all files in a course.
}
\description{
List all files in a course
}
