% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{list_course_rubrics}
\alias{list_course_rubrics}
\title{List Rubrics in course}
\usage{
list_course_rubrics(course_id)
}
\arguments{
\item{course_id}{the canvas id of the course_id (integer)}
}
\value{
dataframe with rubrics info
}
\description{
List all rubrics in a course
}
