% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint_functions.R
\name{push_blueprint}
\alias{push_blueprint}
\title{Push Blueprint}
\usage{
push_blueprint(
  course_id,
  template_id = "default",
  comment = NULL,
  send_notification = F,
  copy_settings = F,
  publish_after_initial_sync = F
)
}
\arguments{
\item{course_id}{the canvas id of the course_id (integer)}

\item{template_id}{template id of the blueprint use "default" as template_id}

\item{comment}{an optional comment to be included in the sync history (string)}

\item{send_notification}{Send a notification to the calling user when the sync completes (boolean)}

\item{copy_settings}{Whether course settings should be copied over to associated courses. Defaults to true for newly associated courses (boolean)}

\item{publish_after_initial_sync}{If set, newly associated courses will be automatically published after the sync completes (boolean)}
}
\value{
server response
}
\description{
Push blueprint pages to associated courses
}
