% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections_to_groups.R
\name{sections_to_groups}
\alias{sections_to_groups}
\title{Sections to Groups}
\usage{
sections_to_groups(course_id, groupcat_name, skip_section_ids = NULL)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{groupcat_name}{the name of the group category (string)}

\item{skip_section_ids}{the canvas id's of sections to skip in copying}
}
\value{
a paginated list with responses
}
\description{
This function turns sections in to group in a Canvas course
}
\examples{
sections_to_groups(course_id = 12345, groupcat_name = "Werkgroepen", skip_section_ids = c(123123, 123124, 123125))
}
