% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{set_course_dates}
\alias{set_course_dates}
\title{Set courses dates}
\usage{
set_course_dates(course_id, start_date = NULL, end_date = NULL)

set_course_dates(course_id, start_date = NULL, end_date = NULL)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{start_date}{starting date in ISO DateTime format or use NA for no date (character)}

\item{end_date}{end date in ISO DateTime format or use NA for no date (character)}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.

server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
This function sets course starting date and end date

This function sets course starting date and end date
}
\examples{

set_course_dates(404, start_date = "2019-01-01T00:00Z")
set_course_dates(404, start_date = NA, end_date = NA )

set_course_dates(404, start_date = "2019-01-01T00:00Z")
set_course_dates(404, start_date = NA, end_date = NA )
}
