% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{set_course_name}
\alias{set_course_name}
\title{Set course name}
\usage{
set_course_name(course_id, name)

set_course_name(course_id, name)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{name}{the new name of the course (character)}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.

server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
This function sets the course name

This function sets the course name
}
\examples{

set_course_name(404, name = "My course name")

set_course_name(404, name = "My course name")
}
