% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{set_front_page}
\alias{set_front_page}
\title{Set front page}
\usage{
set_front_page(course_id, link = NULL, front_page = NULL)

set_front_page(course_id, link = NULL, front_page = NULL)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{link}{url link/ name of the page (character)}

\item{front_page}{make this page front page (boolean)}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.

server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
Set front page

Set front page
}
