% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R
\name{set_grade_on_assignment}
\alias{set_grade_on_assignment}
\title{Set grade}
\usage{
set_grade_on_assignment(
  course_id,
  assignment_id,
  user_id,
  grade = NULL,
  excuse = F
)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{assignment_id}{the canvas id of the assignment (integer)}

\item{user_id}{canvas user id (integer)}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
Conclude deactivate or delete an enrollment
}
