% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{show_page_content_in_course}
\alias{show_page_content_in_course}
\title{Show page content of a course}
\usage{
show_page_content_in_course(course_id, link)

show_page_content_in_course(course_id, link)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{link}{name of the page (character)}
}
\value{
A dataframe containing page content in html format.

A dataframe containing page content in html format.
}
\description{
Gets page content of a course.

Gets page content of a course.
}
\examples{


}
