% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{`API Calls`}
\alias{`API Calls`}
\title{\verb{API Calls}
The functions here are all close to the basic calls in canvas. They can be easily recycled in more complicated functions later down the line.
The plan is to keep extending these functions over time.}
\description{
\verb{API Calls}
The functions here are all close to the basic calls in canvas. They can be easily recycled in more complicated functions later down the line.
The plan is to keep extending these functions over time.
}
