% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint_functions.R
\name{update_blueprint_associations}
\alias{update_blueprint_associations}
\title{Update associated blueprint courses}
\usage{
update_blueprint_associations(
  course_id,
  template_id = "default",
  course_ids_to_add = NULL,
  course_ids_to_remove = NULL
)
}
\arguments{
\item{course_id}{the canvas id of the course_id (integer)}

\item{template_id}{template id of the blueprint use "default" as template_id}

\item{course_ids_to_add}{course id's to add to associations (vector containing integers)}

\item{course_ids_to_remove}{course id's to remove from associations (vector containing integers)}
}
\value{
server response
}
\description{
Updates blueprint associations
}
