% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{update_course}
\alias{update_course}
\title{Update a Course}
\usage{
update_course(course_id, syllabus_body = NULL)
}
\arguments{
\item{syllabus_body}{}
}
\description{
This function allows for changes to all important settings of a course.
}
