% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{update_course_settings}
\alias{update_course_settings}
\title{Update course settings}
\usage{
update_course_settings(
  course_id,
  allow_student_discussion_topics = NULL,
  allow_student_forum_attachments = NULL,
  allow_student_discussion_editing = NULL,
  allow_student_organized_groups = NULL,
  filter_speed_grader_by_student_group = NULL,
  hide_final_grades = NULL,
  hide_distribution_graphs = NULL,
  lock_all_announcements = NULL,
  restrict_student_past_view = NULL,
  restrict_student_future_view = NULL,
  show_announcements_on_home_page = NULL,
  home_page_announcement_limit = NULL
)

update_course_settings(
  course_id,
  allow_student_discussion_topics = NULL,
  allow_student_forum_attachments = NULL,
  allow_student_discussion_editing = NULL,
  allow_student_organized_groups = NULL,
  filter_speed_grader_by_student_group = NULL,
  hide_final_grades = NULL,
  hide_distribution_graphs = NULL,
  lock_all_announcements = NULL,
  restrict_student_past_view = NULL,
  restrict_student_future_view = NULL,
  show_announcements_on_home_page = NULL,
  home_page_announcement_limit = NULL
)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{allow_student_discussion_topics}{allow students to create discussion topics (boolean)}

\item{allow_student_forum_attachments}{allow students to attach files to forums (boolean)}

\item{allow_student_discussion_editing}{allow students to edit discussions (boolean)}

\item{allow_student_organized_groups}{allow students to create organize groups (boolean)}

\item{filter_speed_grader_by_student_group}{filter speed graders by student group (boolean)}

\item{hide_final_grades}{hide grade totals from students (boolean)}

\item{restrict_student_past_view}{Restrict students from viewing courses after end date (boolean)}

\item{restrict_student_future_view}{Restrict students from viewing courses before start date (boolean)}

\item{show_announcements_on_home_page}{show announcements on homepage (boolean)}

\item{home_page_announcement_limit}{set announcements limit on homepage announcements (integer)}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.

server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
This functions update the settings of a course.

This functions update the settings of a course.
}
\examples{

update_course_settings(6348, hide_final_grades = T, hide_distribution_graphs = T)

update_course_settings(6348, hide_final_grades = T, hide_distribution_graphs = T)
}
