% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{update_course_tabs}
\alias{update_course_tabs}
\title{Update course tabs}
\usage{
update_course_tabs(course_id, tab_id, position, hidden)

update_course_tabs(course_id, tab_id, position, hidden)
}
\arguments{
\item{course_id}{the canvas id of the course_id (integer)}

\item{tab_id}{the canvas id of the tab (integer)}

\item{position}{the position of the tab in the course (integer)}

\item{hidden}{make tab hidden or not (boolean)}
}
\value{
a paginated list

a paginated list
}
\description{
Update tabs from course

Update tabs from course
}
