% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{update_page}
\alias{update_page}
\title{Update page content}
\usage{
update_page(
  course_id,
  link = NULL,
  page_title = NULL,
  page_body = NULL,
  front_page = NULL
)

update_page(
  course_id,
  link = NULL,
  page_title = NULL,
  page_body = NULL,
  front_page = NULL
)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{link}{url link/ name of the page (character)}

\item{page_title}{new text written in html (character)}

\item{page_body}{new text written in html (character)}

\item{front_page}{make this page front page (boolean)}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.

server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
This page updates the page content of a page. Content must be written in html.

This page updates the page content of a page. Content must be written in html.
}
