% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{update_page_body}
\alias{update_page_body}
\title{Update page content}
\usage{
update_page_body(course_id, link, page_body)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{link}{url link of the page (character)}

\item{page_body}{new text written in html (character)}
}
\description{
This page updates the page content of a page. Content must be written in html.
}
