% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{create_course_group_category}
\alias{create_course_group_category}
\title{create_course_group_category}
\usage{
create_course_group_category(
  course_id,
  groupcat_name,
  self_signup = NULL,
  auto_leader = NULL,
  group_limit = NULL,
  create_group_count = NULL
)

create_course_group_category(
  course_id,
  groupcat_name,
  self_signup = NULL,
  auto_leader = NULL,
  group_limit = NULL,
  create_group_count = NULL
)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{groupcat_name}{the name of the group category (character/string)}

\item{self_signup}{allow self signup by studentes (boolean)}

\item{auto_leader}{automatically assign leader (boolean)}

\item{group_limit}{set group limit (integer)}

\item{create_group_count}{}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.

server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
create_course_group_category

create_course_group_category
}
\examples{

create_course_group_category(course_id = 6824, groupcat_name = "Werkgroepen")

create_course_group_category(course_id = 6824, groupcat_name = "Werkgroepen")
}
