% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{create_course_group_in_category}
\alias{create_course_group_in_category}
\title{create_course_group_in_category}
\usage{
create_course_group_in_category(groupcat_id, name, description = NULL,
  is_public = NULL, join_level = NULL, storage_quota_mb = NULL,
  sis_group_id = NULL)
}
\arguments{
\item{sis_group_id}{}
}
\description{
create_course_group_in_category
}
