% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{create_course_group_in_category}
\alias{create_course_group_in_category}
\title{create_course_group_in_category}
\usage{
create_course_group_in_category(
  groupcat_id,
  name,
  description = NULL,
  is_public = NULL,
  join_level = NULL,
  storage_quota_mb = NULL,
  sis_group_id = NULL
)

create_course_group_in_category(
  groupcat_id,
  name,
  description = NULL,
  is_public = NULL,
  join_level = NULL,
  storage_quota_mb = NULL,
  sis_group_id = NULL
)
}
\arguments{
\item{groupcat_id}{the canvas id of the group category (integer)}

\item{name}{the name of the group(string/character)}

\item{description}{a description of the group (string/character)}

\item{is_public}{make the group public (boolean)}

\item{storage_quota_mb}{storage quota of the group page in megabytes (integer)}

\item{sis_group_id}{the sis id of the group (integer)}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.

server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
Creates a group in a predefined Canvas group category

Creates a group in a predefined Canvas group category
}
\examples{

create_course_group_in_category(6348, "Werkgroepen_1")
create_course_group_in_category(groupcat_id = 6348, name = "Werkgroepen_1", is_public = T, storage_quota_mb = 20)

create_course_group_in_category(6348, "Werkgroepen_1")
create_course_group_in_category(groupcat_id = 6348, name = "Werkgroepen_1", is_public = T, storage_quota_mb = 20)
}
