% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{create_group_membership}
\alias{create_group_membership}
\title{create_group_membership}
\usage{
create_group_membership(group_id, user_id)

create_group_membership(group_id, user_id)
}
\arguments{
\item{group_id}{the group id (integer)}

\item{user_id}{the canvas id of the user (string). Alternatively, the sis_id can
be used by setting "sis_login_id:" in front of the sis_id.}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.

server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
This function is used to create a group membership for a specific user in 
a specific course.

This function is used to create a group membership for a specific user in 
a specific course.
}
\examples{

create_group_membership(group_id = 4095, user_id = "sis_login_id: pbosman")


create_group_membership(group_id = 4095, user_id = "sis_login_id: pbosman")

}
