% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{delete_course_module}
\alias{delete_course_module}
\title{delete_course_module
Deletes a specified module in a course.}
\usage{
delete_course_module(course_id, module_id)

delete_course_module(course_id, module_id)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{module_id}{the canvas id of the module (integer)}
}
\value{
server response. Either 200 status code if everything went correctly or a specific http status warning.

server response. Either 200 status code if everything went correctly or a specific http status warning.
}
\description{
delete_course_module
Deletes a specified module in a course.

delete_course_module
Deletes a specified module in a course.
}
\examples{

delete_course_module(course_id = 6824, module_id = 412)

delete_course_module(course_id = 6824, module_id = 412)
}
