% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{get_assignment_stats_in_course}
\alias{get_assignment_stats_in_course}
\title{Get assignment analytics}
\usage{
get_assignment_stats_in_course(course_id)

get_assignment_stats_in_course(course_id)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}
}
\value{
A dataframe containing assignment statistics

A dataframe containing assignment statistics
}
\description{
Get assignment statistics in a course

Get assignment statistics in a course
}
\examples{

get_assignments_stats_in_course(6824)

get_assignments_stats_in_course(6824)
}
