% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_assignment_submissions}
\alias{get_assignment_submissions}
\title{Get assignment submissions}
\usage{
get_assignment_submissions(course_id, assignment_id, include = NULL)
}
\arguments{
\item{course_id}{the canvas id of the course_id (integer)}

\item{assignment_id}{the canvas id of the assignment (integer)}

\item{include}{rubric_assessment}
}
\value{
a paginated list
}
\description{
Get all submission info on a assignment
}
