% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_groups_in_course}
\alias{get_groups_in_course}
\title{Get groups in course}
\usage{
get_groups_in_course(course_id)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}
}
\value{
A dataframe containing information about all groups in a course
}
\description{
This function returns data about all groups in a specified course
}
\examples{

get_groups_in_course(6348)
}
