% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{get_members_in_group}
\alias{get_members_in_group}
\title{Get members in group}
\usage{
get_members_in_group(group_id)
}
\arguments{
\item{group_id}{the canvas id of the group (integer)}
}
\value{
A dataframe with information about members of a specified group
}
\description{
This functions gets user info of a group membership
}
\examples{

get_members_in_group(11267)
}
