% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{get_section_enrollments}
\alias{get_section_enrollments}
\title{get_section_enrollments}
\usage{
get_section_enrollments(section_id)

get_section_enrollments(section_id)
}
\arguments{
\item{section_id}{the canvas id of the section (integer)}
}
\value{
a dataframe containing all the section enrollments

a dataframe containing all the section enrollments
}
\description{
This function gets all the enrollments for a specific section in a course.

This function gets all the enrollments for a specific section in a course.
}
\examples{

get_section_enrollments(17234)


get_section_enrollments(17234)

}
