% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{get_single_assignment_in_course}
\alias{get_single_assignment_in_course}
\title{Get single assignment in course}
\usage{
get_single_assignment_in_course(course_id, assignment_id)

get_single_assignment_in_course(course_id, assignment_id)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{assignment_id}{the canvas id of the assignment (integer)}
}
\value{
A dataframe containing information about a specified assignment from a specified course.

A dataframe containing information about a specified assignment from a specified course.
}
\description{
Get the information of a single assignment in a course

Get the information of a single assignment in a course
}
\examples{

get_single_assingment_in_course(6348, 10234)

get_single_assingment_in_course(6348, 10234)
}
