% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{get_user_details}
\alias{get_user_details}
\title{Get user details}
\usage{
get_user_details(user_id, include = NULL)

get_user_details(user_id, include = NULL)
}
\arguments{
\item{user_id}{The canvas id of the user(integer)}

\item{include}{Include arguments: "email"}
}
\value{



}
\description{
Get information about a user

Get information about a user
}
