% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{get_users_in_account}
\alias{get_users_in_account}
\title{Get users in account}
\usage{
get_users_in_account(account_id, enrollment_type = NULL)

get_users_in_account(account_id, enrollment_type = NULL)
}
\arguments{
\item{account_id}{the canvas id of the account (integer)}

\item{enrollment_type}{the canvas enrollment type. Use either "student", "teacher", "ta", "observer" or "designer" (character)}
}
\value{
A dataframe containing information about users in the specified subaccount

A dataframe containing information about users in the specified subaccount
}
\description{
Gets all user information from a specified account. You can use the argument enrollment_type to set a specific filter to user type, to save some time.

Gets all user information from a specified account. You can use the argument enrollment_type to set a specific filter to user type, to save some time.
}
\examples{

get_users_in_account(123, "teacher")

get_users_in_account(123, "teacher")
}
