% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions (conflicted copy 2021-04-14
%   155304).R, R/simple_functions.R
\name{get_users_in_course}
\alias{get_users_in_course}
\title{Get users in course}
\usage{
get_users_in_course(
  course_id,
  enrollment_type = NULL,
  enrollment_state = NULL,
  include = "enrollments"
)

get_users_in_course(
  course_id,
  enrollment_type = NULL,
  enrollment_state = NULL,
  include = "enrollments"
)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{enrollment_type}{the canvas enrollment type to return. Use either "StudentEnrollment", "TeacherEnrollment". Default returns all roles. (character)}

\item{enrollment_state}{include users with a specific state. Use either "active", "invited", "deleted" or "completed". Default is "active"  and "invited" (character)}
}
\value{
A dataframe containing information about users in the specified course.

A dataframe containing information about users in the specified course.
}
\description{
Gets all user information from a specified course. You can use the argument enrollment_type to set a specific filter to user type 
or the argument enrollment_state for a specific enrollment state of the user, to save some time.

Gets all user information from a specified course. You can use the argument enrollment_type to set a specific filter to user type 
or the argument enrollment_state for a specific enrollment state of the user, to save some time.
}
\examples{

get_users_in_course(6348, enrollment_type = "student", enrollment_state = "active")

get_users_in_course(6348, enrollment_type = "student", enrollment_state = "active")
}
