% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{update_course}
\alias{update_course}
\title{Update a Course}
\usage{
update_course(course_id, syllabus_body = NULL, default_view = NULL)
}
\arguments{
\item{course_id}{the canvas id of the course (integer)}

\item{syllabus_body}{the syllabus text body written in html (character)}

\item{default_view}{change default view "modules" or "wiki"(character)}
}
\description{
This function allows for changes to the syllabus body.
}
